# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "astro API",
    "version": "1.0.0",
    "description": "Threat intelligence search across MITRE ATT&CK, VirusTotal, abuse.ch, AbuseIPDB, OTX, GreyNoise, Shodan, NVD, CISA KEV, EPSS and the IEEE MAC registry. Online sources receive the searched indicators: use `offline` for sensitive ones.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/providers": {
      "get": {
        "summary": "List intelligence sources",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Source"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "summary": "Search one indicator",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 2048
            },
            "description": "Indicator: hash, IP, domain, URL, email, CVE, ATT&CK ID, MAC or a name. Defanged input is accepted."
          },
          {
            "name": "only",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated source names."
          },
          {
            "name": "offline",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "no_cache",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "post": {
        "summary": "Search several indicators",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "indicators"
                ],
                "additionalProperties": false,
                "properties": {
                  "indicators": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 100,
                    "items": {
                      "type": "string"
                    }
                  },
                  "only": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Query only these sources (see /providers)."
                  },
                  "offline": {
                    "type": "boolean",
                    "description": "Use only offline datasets: the indicator never leaves the server."
                  },
                  "no_cache": {
                    "type": "boolean",
                    "description": "Ignore cached results."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Report"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/extract": {
      "post": {
        "summary": "Extract indicators from text without searching",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "additionalProperties": false,
                "properties": {
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "indicators": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Indicator"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/enrich": {
      "post": {
        "summary": "Extract indicators from an alert or text and search them",
        "description": "Designed for SOAR playbooks (e.g. SentinelOne Singularity Hyperautomation): send the alert as text, branch on `verdict`. At most 50 indicators are searched.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "indicators": {
                    "type": "array",
                    "maxItems": 100,
                    "items": {
                      "type": "string"
                    }
                  },
                  "only": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Query only these sources (see /providers)."
                  },
                  "offline": {
                    "type": "boolean",
                    "description": "Use only offline datasets: the indicator never leaves the server."
                  },
                  "no_cache": {
                    "type": "boolean",
                    "description": "Ignore cached results."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token created with `astro token create <name>`."
      }
    },
    "responses": {
      "Error": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, invalid or revoked token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Verdict": {
        "type": "string",
        "enum": [
          "",
          "info",
          "clean",
          "suspicious",
          "malicious"
        ]
      },
      "Indicator": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "md5",
              "sha1",
              "sha256",
              "sha512",
              "ipv4",
              "ipv6",
              "domain",
              "url",
              "email",
              "cve",
              "attack-technique",
              "attack-tactic",
              "attack-group",
              "attack-software",
              "attack-mitigation",
              "attack-campaign",
              "mac",
              "keyword"
            ]
          },
          "value": {
            "type": "string"
          }
        }
      },
      "Field": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "Result": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "found": {
            "type": "boolean"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "summary": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Field"
            }
          },
          "reference": {
            "type": "string",
            "format": "uri"
          },
          "details": {
            "description": "Source-specific structured data"
          },
          "error": {
            "type": "string"
          },
          "cached": {
            "type": "boolean"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Report": {
        "type": "object",
        "properties": {
          "indicator": {
            "$ref": "#/components/schemas/Indicator"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Result"
            }
          },
          "started": {
            "type": "string",
            "format": "date-time"
          },
          "duration_ns": {
            "type": "integer"
          }
        }
      },
      "Source": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "offline": {
            "type": "boolean"
          },
          "key_env": {
            "type": "string"
          },
          "indicators": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "EnrichResponse": {
        "type": "object",
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "searched": {
            "type": "integer"
          },
          "truncated": {
            "type": "boolean"
          },
          "malicious": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Indicator"
            }
          },
          "suspicious": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Indicator"
            }
          },
          "reports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Report"
            }
          }
        }
      },
...
    }
  }
}